优化输出

+++ note/0002_进程和线程/test.cc
@@ -9,9 +9,12 @@ std::map<int, std::string> m;
 
 void handle_signal(int sig, siginfo_t* sig_info, void * ) {
     std::cout << getpid() << " 捕获信号 " << m[sig] << " 来自 " << sig_info->si_pid << std::endl;
-//    std::cout << getpid() << " 处理信号 " << m[sig] << " 中..." << std::endl;
-//    sleep(2);
-//    std::cout << getpid() << " 处理完成 " << m[sig] << std::endl;
+    // 对 SIGUSR1 SIGUSR2 特殊处理
+    if (sig == SIGUSR1 || sig == SIGUSR2) {
+        std::cout << getpid() << " 处理信号 " << m[sig] << " 中..." << std::endl;
+        sleep(2);
+        std::cout << getpid() << " 处理完成 " << m[sig] << std::endl;
+    }
 }
 
 void set_signal(bool block = false) {
@@ -29,7 +32,6 @@ void set_signal(bool block = false) {
 
 void send_signal(int sig, pid_t fd) {
     kill(fd, sig);
-    //std::cout << getpid() << " 发送信号 " << m[sig] << std::endl;
 }
 
 void init() {
@@ -71,21 +73,23 @@ int main() {
 
 #if 0
     // 测试信号处理函数执行过程中, 相同的信号到达多个的情况
+    // 结论: 信号处理过程中, 相同信号到来会阻塞, 多个相同信号后续只处理一个, 多余的会丢弃
     pid_t fd = fork();
 
     if (fd == 0) {
-        std::cout << "子进程: " << getpid() << std::endl;
         // 子进程
         set_signal();
+        std::cout << getpid() << " 设置信号处理函数完成" << std::endl;
         for (;;)
             ;
     } else {
         // 父进程
         sleep(2); // 保证子进程已设置信号处理
-        std::cout << "父进程: " << getpid() << std::endl;
+        std::cout << getpid() << " 发送信号 " << m[SIGUSR1] << std::endl;
         send_signal(SIGUSR1, fd);
         sleep(1);
-        for (int i = 1; i < 11; ++i) {
+        for (int i = 1; i < 5; ++i) {
+            std::cout << getpid() << " 发送信号 " << m[SIGUSR1] << std::endl;
             send_signal(SIGUSR1, fd);
         }
         for (;;)
@@ -95,21 +99,24 @@ int main() {
 
 #if 0
     // 测试信号处理函数执行过程中, 不同的信号到达多个的情况
+    // 结论: 信号处理过程中, 不同信号到来会直接处理, 多个不同信号后续只处理一个, 多余的会丢弃
+    //       处理完成以后, 再处理之前的信号
     pid_t fd = fork();
 
     if (fd == 0) {
-        std::cout << "子进程: " << getpid() << std::endl;
         // 子进程
         set_signal();
+        std::cout << getpid() << " 设置信号处理函数完成" << std::endl;
         for (;;)
             ;
     } else {
         // 父进程
         sleep(2); // 保证子进程已设置信号处理
-        std::cout << "父进程: " << getpid() << std::endl;
+        std::cout << getpid() << " 发送信号 " << m[SIGUSR1] << std::endl;
         send_signal(SIGUSR1, fd);
         sleep(1);
-        for (int i = 1; i < 11; ++i) {
+        for (int i = 1; i < 5; ++i) {
+            std::cout << getpid() << " 发送信号 " << m[SIGUSR2] << std::endl;
             send_signal(SIGUSR2, fd);
         }
         for (;;)
@@ -127,7 +134,7 @@ int main() {
     sigprocmask(SIG_SETMASK, &mask, NULL);
     std::cout << "开始发送信号" << std::endl;
     for (auto key : m)
-        if (key.first != SIGKILL && key.first != SIGSTOP)
+        if (key.first != SIGKILL && key.first != SIGSTOP && key.first != SIGUSR1 && key.first != SIGUSR2)
             send_signal(key.first, getpid());
     std::cout << "发送信号完成" << std::endl;
     std::cout << "解除信号阻塞" << std::endl;
@@ -138,16 +145,18 @@ int main() {
 #endif
 
 #if 0
+    // 测试 core
     int x = 0;
-
-    std::cout << 1/ x  << std::endl;
+    std::cout << 1 / x  << std::endl;
 #endif
 
+#if 1
     set_signal();
     std::cout << "父进程: " << getppid() << std::endl;
     std::cout << "  进程: " << getpid()  << std::endl;
     for (;;)
         ;
+#endif
 
     return 0;
 }
